+++ Kaidel/src/Platform/Vulkan/GraphicsAPI/VulkanShader.h
@@ -9,7 +9,7 @@ namespace Kaidel {
 		~VulkanShader();
 
 		auto& GetShaderInfo()const { return m_Shader; }
-		virtual uint32_t GetSetCount()const override { return m_Shader.DescriptorSetLayouts.size(); }
+		virtual uint32_t GetSetCount()const override { return (uint32_t)m_Shader.DescriptorSetLayouts.size(); }
 
 	private:
 		VulkanBackend::ShaderInfo m_Shader;
